+++ src/main/java/nl/utwente/groove/explore/Generator.java
@@ -427,19 +427,19 @@ public class Generator extends GrooveCmdLineTool<ExploreResult> {
     /** Usage message for the strategy option. */
     public final static String STRATEGY_USAGE
         = "Set the exploration strategy to <strgy>. Legal values are:\n"
-            + "  bfs[:n]     - Optionally bounded Breadth-First exploration: "
+            + "  bfs[:n]     - Optionally bounded Breadth-First exploration: \n"
             + "                if n>0, exploration stops at depth n\n" //
-            + "  dfs[:n]     - Optionally bounded Depth-First exploration: "
+            + "  dfs[:n]     - Optionally bounded Depth-First exploration: \n"
             + "                if n>0, exploration stops at depth n\n" //
             + "  linear      - Linear\n" //
             + "  random      - Random linear\n" //
             + "  uptorule:[dfs|bds][->|=>][!]id\n" //
-            + "              - BFS or DFS up to (for ->) or including (for =>) states "
+            + "              - BFS or DFS up to (for ->) or including (for =>) states \n"
             + "                where rule <id> is or is not (!) applicable\n" //
-            + "  cnbound:n   - BFS up to (but not including) graphs with "
+            + "  cnbound:n   - BFS up to (but not including) graphs with \n"
             + "                more than <n> nodes\n" //
             + "  cebound:id_1>n_1,...,id_k>n_k\n" //
-            + "              - BFS up to (but not including) graphs with "
+            + "              - BFS up to (but not including) graphs with \n"
             + "                more than <n_i> <id_i>-edges, for all i in 1..k\n"
             + "  ltl:prop    - LTL Model Checking\n" //
             + "  ltlbounded:idn,...;prop\n" //