+++ src/main/java/nl/utwente/groove/explore/strategy/BoundedLTLStrategy.java
@@ -1,17 +1,17 @@
 /*
  * GROOVE: GRaphs for Object Oriented VErification Copyright 2003--2023
  * University of Twente
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
- * 
+ *
  * $Id: DefaultBoundedModelCheckingStrategy.java,v 1.1 2008/03/04 14:44:25
  * kastenberg Exp $
  */
@@ -26,7 +26,7 @@ import nl.utwente.groove.verify.ProductTransition;
 /**
  * This class provides some default implementations for a bounded model checking
  * strategy, such as setting the boundary and collecting the boundary graphs.
- * 
+ *
  * @author Harmen Kastenberg
  * @version $Revision$
  */
@@ -153,18 +153,16 @@ public class BoundedLTLStrategy extends LTLStrategy {
      *         blue, or red, <tt>false</tt> otherwise
      */
     protected boolean isUnexplored(ProductState newState) {
-        boolean result =
-            newState.colour() != getRecord().cyan() && newState.colour() != getRecord().blue()
-                && newState.colour() != getRecord().red();
+        boolean result = newState.colour() != getRecord().cyan()
+            && newState.colour() != getRecord().blue() && newState.colour() != getRecord().red();
         return result;
     }
 
     @Override
-    protected ProductState getNextSuccessor(ProductState state) {
-        ProductState result = null;
+    protected ProductTransition getNextSuccessor(ProductState state) {
+        ProductTransition result = null;
         // pick a transition to an unexplored state
-        RandomChooserInSequence<ProductTransition> chooser =
-            new RandomChooserInSequence<>();
+        RandomChooserInSequence<ProductTransition> chooser = new RandomChooserInSequence<>();
         for (ProductTransition p : state.outTransitions()) {
             ProductState buchiState = p.target();
             if (isUnexplored(buchiState)) {
@@ -181,7 +179,7 @@ public class BoundedLTLStrategy extends LTLStrategy {
             // the current depth of the boundary should be updated
             getBoundary().crossingBoundary(resultTransition, true);
             // and the state reached by that transition
-            result = resultTransition.target();
+            result = resultTransition;
         }
         return result;
     }