+++ src/main/java/nl/utwente/groove/grammar/GrammarKey.java
@@ -154,11 +154,6 @@ public enum GrammarKey implements Properties.Key, GrammarChecker {
     /**
      * Space-separated list of disabled rule names.
      */
-    DISABLED_RULES("disabledRules", "List of explicitly disabled rules", ValueType.QUAL_NAME_LIST),
-
-    /**
-     * Space-separated list of disabled rule names.
-     */
     RULE_ENABLING("ruleEnabling",
         "List of force-enabled and -disabled rules (prefixed '+' or '-', respectively).<br>"
             + "This overrides the 'Enabled' rule property.",
@@ -309,7 +304,7 @@ public enum GrammarKey implements Properties.Key, GrammarChecker {
             case COMMON_LABELS, CONTROL_LABELS -> Parser.splitter;
             case CREATOR_EDGE, PARALLEL, DANGLING, RHS_AS_NAC, INJECTIVE, STORE_OUT_PARS, USE_STORED_NODE_IDS -> Parser.boolFalse;
             case ISOMORPHISM, LOOPS_AS_LABELS -> Parser.boolTrue;
-            case START_GRAPH_NAMES, CONTROL_NAMES, TYPE_NAMES, PROLOG_NAMES, DISABLED_RULES -> QualName
+            case START_GRAPH_NAMES, CONTROL_NAMES, TYPE_NAMES, PROLOG_NAMES -> QualName
                 .listParser();
             case RULE_ENABLING -> DeltaMap.parser(QualName.parser());
             case TYPE_POLICY -> new Parser.EnumParser<>(CheckPolicy.class, CheckPolicy.ERROR,
@@ -356,7 +351,6 @@ public enum GrammarKey implements Properties.Key, GrammarChecker {
         case PROLOG_NAMES -> ResourceChecker.get(ResourceKind.PROLOG);
         case START_GRAPH_NAMES -> ResourceChecker.get(ResourceKind.HOST);
         case TYPE_NAMES -> ResourceChecker.get(ResourceKind.TYPE);
-        case DISABLED_RULES -> ResourceChecker.get(ResourceKind.RULE);
         case RULE_ENABLING -> RuleDeltaChecker.instance();
         case USER_OPS -> UserOperationsChecker.instance;
         default -> trueChecker;
@@ -373,7 +367,7 @@ public enum GrammarKey implements Properties.Key, GrammarChecker {
     @Override
     public boolean isNotable() {
         return switch (this) {
-        case ACTION_POLICY, ALGEBRA, CREATOR_EDGE, DANGLING, DEAD_POLICY, INJECTIVE, ISOMORPHISM, ORACLE, RHS_AS_NAC, STORE_OUT_PARS, TRANSITION_PARAMETERS, RULE_ENABLING, DISABLED_RULES, TYPE_POLICY -> true;
+        case ACTION_POLICY, ALGEBRA, CREATOR_EDGE, DANGLING, DEAD_POLICY, INJECTIVE, ISOMORPHISM, ORACLE, RHS_AS_NAC, STORE_OUT_PARS, TRANSITION_PARAMETERS, RULE_ENABLING, TYPE_POLICY -> true;
         default -> false;
         };
     }